Refactor:moved all buton in app/components/Buttons/ and changed all imports in diferents files
